Move Highlighter to its own file
